+++ root/programs/MultiPurposeAuthSite/MultiPurposeAuthSite/MultiPurposeAuthSite.csproj
@@ -250,7 +250,8 @@
     <Compile Include="Controllers\ErrorController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
     <Compile Include="Controllers\ManageController.cs" />
-    <Compile Include="Controllers\OAuth2ResourceApiController.cs" />
+    <Compile Include="Controllers\OAuth2EndpointApiController.cs" />
+    <Compile Include="Controllers\OAuth2ResourceServerColler.cs" />
     <Compile Include="Controllers\PingController.cs" />
     <Compile Include="Controllers\RolesAdminController.cs" />
     <Compile Include="Controllers\UsersAdminController.cs" />
